Add recycler view adapter for daily forecast

+++ app/src/main/java/com/grzeluu/weatherapp/model/Daily.kt
@@ -1,7 +1,7 @@
 package com.grzeluu.weatherapp.model
 
 data class Daily(
-    val dt: Int,
+    val dt: Long,
     val sunrise: Long,
     val sunset: Long,
     val moonrise: Long,
